Encoder/Decoder
This encodes/decodes the messages sent/received from the server.  
This module supports sending parts of the message and the message at once, Example you can send the first 2 bytes that represents the operation type (Ack/Follow/...) and only later send the rest of the message, STILL! this module knows that these bytes belong to a previous unfinished message.